+++ src/Api/Dsl/GivenClassOrInterface/AndMethod/AddParam.php
@@ -5,8 +5,12 @@ namespace Fjor\Api\Dsl\GivenClassOrInterface\AndMethod;
 interface AddParam
 {
     /**
-     * Specifies the values that should be injected in a method. `addParam`
-     * can be called more than once to inject other values.
+     * Specifies the values that should be injected in a method.
+     * 
+     * Example: `addParam(2, 4, 8)`
+     * 
+     * `addParam`can be called more than once to inject other values. The
+     * method then gets called multiple times.
      */
-    public function addParam(array $values = array()) : AddParam;
+    public function addParam(...$values) : AddParam;
 }